refactor natcrit icon logo into a comp
