+++ client/naturalcrit/accountPage/accountPage.jsx
@@ -1,17 +1,11 @@
 import React from 'react';
 const AccountActions = require('../account.actions.js');
-const NaturalCritIcon = require('naturalcrit/svg/naturalcrit.svg.jsx');
+const NaturalCritIcon = require('naturalcrit/components/naturalcritLogo.jsx');
 
 const AccountPage = (props) => {
 	return (
 		<div className="accountPage">
-			<a className='logo' href='/'>
-				<NaturalCritIcon />
-				<span className='name'>
-					Natural
-					<span className='crit'>Crit</span>
-				</span>
-			</a>
+			<NaturalCritIcon />
 			<div className="details">
 				<h1>Account Page</h1>
 				<br />
